requirements.txt
astroid==1.6.4
certifi==2018.8.13
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-decouple==2.1
django-import-export==1.0.1
django-js-asset==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==3.2.3
et-xmlfile==1.0.1
idna==2.7
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.5
Pillow==5.2.0
psycopg2==2.7.5
PyJWT==1.6.4
pylint==1.9.1
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
